refactor: atomic pattern으로 컴포넌트 정리
