+++ src/pages/account/SignUp.tsx
@@ -6,8 +6,8 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
-import PasswordFormField from '@/components/PasswordFormField';
-import EmailFormField from '@/components/EmailFormField';
+import PasswordFormField from '@/components/molecule/field/PasswordFormField';
+import EmailFormField from '@/components/molecule/field/EmailFormField';
 import { AppDispatch } from '@/RTK/store';
 import { signUp } from '@/RTK/thunk';
 import { useDispatch } from 'react-redux';
